+++ faf-java-server-app/src/main/java/com/faforever/server/ice/IceServersProvider.java
@@ -1,5 +1,11 @@
 package com.faforever.server.ice;
 
+
+import com.faforever.server.player.Player;
+
+/**
+ * Interface for ICE server list providers.
+ */
 public interface IceServersProvider {
-  IceServerList getIceServerList();
+  IceServerList getIceServerList(Player player);
 }