+++ changelog.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Custom ICE server credentials now use long term credentials as specified in RFC 5389
 
 ### Fixed
-
 - Don't require login for actuator endpoints
 
 ## v1.3.4